requirements.txt
fastapi
uvicorn
gunicorn
line-bot-sdk
firebase-admin
google-generativeai
pytz
requests
python-multipart
python-dotenv
pytest
pytest-mock